refactor: Correct import path for `utils` module
