+++ lib/error.js
@@ -8,7 +8,7 @@
 
 'use strict';
 
-const { isPlainObject } = require('./utils');
+const { isPlainObject } = require('./utils/type-utils');
 
 /**
  * @classdesc Represents an error that occurred during video ID extraction.
