+++ lib/utils/url-utils.js
@@ -10,8 +10,8 @@
 
 'use strict';
 
-const { isNullOrUndefined } = require('./utils');
-const { IDExtractorError } = require('./error');
+const { isNullOrUndefined } = require('./type-utils');
+const { IDExtractorError } = require('../error');
 
 
 /**
