+++ lib/ytmp3.js
@@ -55,6 +55,7 @@ const {
   LOGDIR,
   logger: log,
   ProgressBar,
+  URLUtils,
   createDirIfNotExist,
   createDirIfNotExistSync,
   createLogFile
@@ -64,7 +65,6 @@ const {
   convertAudio,
   defaultOptions: defaultAudioConvOptions
 } = require('./audioconv');
-const URLUtils = require('./url-utils');
 
 /**
  * The video information object.
